Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> @Luigi/Jesse - i would be against reverting back to opendev, since the instability of it seemed too big causing lot of failures in CI, plus the additional 5x slower then github.

A simple approach with only the cost of slower speed, is https://review.gerrithub.io/c/redhat-openstack/infrared/+/490902

> But then yes if github repos are not to be relied on as kept in sync, we need different way to deal with it.

Another approach could be to have the initial clone done from github, then switch to opendev and fetch. However, this adds the cost of complexity and I don't really think it's worth it.

Another approach could be to use a common location for git clones on the CI hosts so that we're only doing updates during every CI run, rather than full clones.

> Most likely will need to add these repos to our internal mirrors (most already there i guess) and make them to be in the up-to-1minute sync repo set. Or maybe come up with another alternative proxy-like (always up-to-date unless source down/unreachable) approach (dunno how well that would work as git repo)?

This sounds like additional complexity and management overhead to me. Not only do the mirrors need to be setup and managed, but infrared and jenkins jobs need to be changed to support overriding the individual git repositories.

> Anyway i suggest if there is really concern with github not having required content to make a internal infra ticket about this? As also maybe check with upstream if the stability and bandwidth of opendev could be improved?

Considering that the openstack infra team handles many, many more tests than we do on a day to day basis, perhaps it'd be worth building a relationship between downstream infra and upstream infra and understanding what the problem is here and how to work together to solve it.

Patch-set: 1