[virsh] Use common create_venv role for swift client
Rather than build the venv our own way, we can now use the
common role which caters for all the crazy things that need
to be done to workaround CI issues.

Due to the way that IR targets the hypervisor the virtualenv
may end up owned by root and therefore the installation of
pypi packages fail to install into it. To counteract this
annoying side-effect, we ensure that the virtualenv is owned
by the same user/group as the parent directory before we
begin to install into it.

Change-Id: I2a2ca809c62f6bfee6c39b05cd7b3576b6d264fa