Make bmc os-cloud parametric
The deployment could fail due to the fact the os-cloud name
is hardcoded into the systemd unit in the BCM instance.
If the cloud name is different than "bmc", the systemd unit
fails and the service isn't available.
This patch just make it parametric, allowing the unit to
properly use the configured cloud and discover the overcloud
instances.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>
Change-Id: Ife13b485a77cbe8bf945f801f1f752a6dd4bd462