Update patch set 7
Patch Set 7: Code-Review+2

> Patch Set 4:
> 
> (1 comment)

Hi Jesse,

The reason this task was written from the first place was because some comopenent/s required the newer version of virtualenv. (I don't remember which one/s and wasn't able to find the ticket)
Itzik tried to run few jobs of different OSP versions without this task and it failed on OSP 16.
I do agree that this task is probably not needed anymore, especially if it was fixed in different (earlier) stages.
Some of his jobs are blocked until this one will be merged and it only moves us one step forward in the direction of completely removing this stage.

Patch-set: 7
Reviewer: Gerrit User 1004057 <1004057@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2