[virsh] Change machine type for compatibility
When saving the xml files from the original libvirt installation,
the machine type will look something like:

<domain type='kvm'>
  <os>
    <type arch='x86_64' machine='pc-i440fx-rhel7.6.0'>hvm</type>
  </os>
</domain>

For each RHEL host, the '-rhel7.6.0' postfix will be different and
it will evolve, however when inspecting the host capabilities it
is evident that this matches the host's canonical machine type for
the hvm guest type.

<capabilities>
  <guest>
    <os_type>hvm</os_type>
    <arch name='x86_64'>
      <machine maxCpus='240'>pc-i440fx-rhel7.6.0</machine>
      <machine canonical='pc-i440fx-rhel7.6.0' maxCpus='240'>pc</machine>
      <domain type='kvm'/>
    </arch>
  </guest>
</capabilities>

Rather than trying to figure out the canonical machine type on the
hypervisor, we rather just change the machine type to 'pc' so that
it uses the canonical type.

Change-Id: Ib31fe0b6db85052d6e29336123225990a9b7ebf6