Remove version-descovery call for upgrades.
The tripleo-overcloud role will always call the version-discovery role
at the very beginning of its execution. For that reason, there is no need
to re-run the version-discovery role again during upgrades as it seems to
interfere with the version calculated at the very begginning and the
overcloud nodes do not get set the proper version from the undercloud.

Change-Id: If243f9b7f2cd1ff5e530df87cce3fc4f91562036