Fix version truncate on overcloud update
We truncate to whole number the undercloud version.
This code shouldn't be handled here since version discovery already
deals with the undercloud version in case the release is lower than 16.x
By doing that we ignore the version discovery logic and might truncate
the version to whole number, for example 16.1|int => 16.

Change-Id: I9af50d21ab23c2e7925e27c9b14f8694866fe4dc