Remove version-discovery call for updates.
The tripleo-overcloud role will always call the version-discovery
role at the very beginning of its execution. For that reason, there
is no need to re-run the version-discovery role again during updates
as it seems to interfere with the version calculated at the very
begginning and the overcloud nodes do not get set the proper version
from the undercloud.

Change-Id: I9af50d21ab23c2e7925e27c9b14f8694866fe4dc