Update to recent hacking
The hacking<2.0.0 causes issues with Python3. The version was pinned
temporarily some time ago to unbreak the gates. I reviewed reported
issues and made a complete list of ignores that can be taken one by
one in the future for resolving. Also I removed non-existing entries
from exclude section.

Basically it reverts change "Temporarily cap hacking to < 2.0.0"
and explicitly lists the issues in [flake8] section of tox.ini file
so we are aware of existing problems, not not just shade them
with old checkers.

This reverts commit 265549e3c0237fd9a8c7733db698ec5891590097.

Change-Id: I231dd0a604766a9c3de6d1400f3afe8b9eae41cd