Allow ansible to determine interpreter for local connections
Ansible can do this itself, so we do not need to set it. It should
default to setting ansible_python_interpreter to the interpreter
provided by ansible_playbook_python. Setting it in inventory
causes it to persist when using the same environment over multiple
jobs, which means that tasks will fail unless the inventory is
updated.

Change-Id: Ied5c78bb9479eba06517f6b9b10241fddb36c863