Fix: 'dict object' has no attribute 'tester'
When we revert to ansible 2.7, these tasks fail with
the error:

The task includes an option with an undefined variable.
The error was: 'dict object' has no attribute 'tester'

This is due to us losing the functionality provided by
ansible 2.8 [a] where nested variables do not require
every attribute in the data structure to be defined.

[a] https://docs.ansible.com/ansible/latest/porting_guides/porting_guide_2.8.html#jinja-undefined-values

Change-Id: I4c3f69f3bd82e70aee438836a964523b335c0183