Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> why not host-pyshical-mtu and host-virtual-mtu?
> Also, it specifies the default is 1500 but the default is set to "False". Is that on purpose? Did you mean the default in the product is 1500?

Hi Arie, the patch is intended only for virt setups i.e. virtual networks and virtual nodes. It is important that they both are configured with same custom MTU and I think there is no use case for configuring 2 different values. There is another parameter with a little bit confising name --host-mtu-size which works only for physical interfaces which is in use by NFV guys as they are working with baremetal computes. Do you mean that it is better to rename my parameter to host-virtual-mtu and host-mtu-size to host-physical-mtu? I think it will cause unnecessary issues with NFV team and maybe others who are using baremetal computes. 

By default the parameter is disabled and the default product 1500 will be used as usual.

Patch-set: 2