Update patch set 16
Patch Set 16:

I think what is wrong with this change is having put so much meat in one single change, so that it has been reverted without any investigation or reporting any of the reason it broke whatever job. We should split changes in smaller pieces and merge them one by one as small pills so that they are not reverted all in one single block. For example probably by only switching to Ansible 2.8 it wouldn't have produce the same failures and gave us simpler errors to fix before applying other improvements to IR plugins.

Patch-set: 16
CC: Gerrit User 1013661 <1013661@d5d70762-12d0-45a1-890d-524b12d3f735>