Update patch set 16
Patch Set 16:

@Federico - It was originally done in https://review.gerrithub.io/c/redhat-openstack/infrared/+/481730 and reverted in https://review.gerrithub.io/c/redhat-openstack/infrared/+/485297

The changes were bundled together with some other things in part to get more testing initiated. This was done at the request of the INFRA folks.

You'll find that RHOSINFRA-3056 covers some history and notes things that broke in this last attempt. A key problem, though, is that there is no infrared gating which covers all plugins, so it's down to manually triggering tests or merging it, failing, reverting, and trying again.

A significant amount of work has gone into this, for infrared much of it is in https://review.gerrithub.io/q/(topic:%2522ansible2.8%2522+OR+topic:%2522py3%2522)+(status:open+OR+status:merged) but the work spans multiple repositories and do not all reflect in gerrithub. There is a lot more to do, and if you're picking that up - great... I won't be able to get to it any time soon.

I would recommend looking at RHOSINFRA-3057 and perhaps tackling that with more urgency because that gives us a more stable interface for using ansible.

Patch-set: 16