Verify TripleO images modification
This change introduces the Ansible playbok that ensures the necessary
containers images modification was performed successfully.

It first checks the presence of /patched_rpms/ directory on the
undercloud node -- if it is there, then the content is registered
as list of packages names and theirs details (version string,
release, etc.).

Then on each overcloud node, in each running container, the installed
packages, with names matching those that should have been patched,
are queried for version strings and the result is compared (diff)
with a list of expected versions. If there is any unexpected version
of matched package in container (entry on the right side of the diff,
i.e. something was not patched), an error is raised.

Jira: RHOSINFRA-3347
Change-Id: I7540b08f99e1f4137199df915dff918791964701