Update patch set 11
Patch Set 11:

(1 comment)

> Patch Set 11:
> 
> And my osp16 failed during the OC install with:
> TASK [ipaclient : Install - Ensure that IPA client packages are installed] *****
> Wednesday 20 May 2020  09:34:44 +0000 (0:00:00.442)       0:32:55.050 *********
> fatal: [networker-0]: FAILED! => {"changed": false, "msg": "No group idm:DL1/client available.", "results": []}
> fatal: [networker-1]: FAILED! => {"changed": false, "msg": "No group idm:DL1/client available.", "results": []}
> fatal: [networker-2]: FAILED! => {"changed": false, "msg": "No group idm:DL1/client available.", "results": []}

That makes sense.  The problem is that ansible-freeipa tries to install the ipa client packages, as defined by the group "idm:DL1/client", but if the system is not registered, then it can't find the group.  Fortunately, the image already has all the packages it needs.

There is a patch upstream to fix this -- https://review.opendev.org/#/c/728441/  which I'm in the process of testing.

Patch-set: 11
CC: Gerrit User 1006809 <1006809@d5d70762-12d0-45a1-890d-524b12d3f735>