[virsh] Implement snapshot flag for CI environments
In CI environments disk space can be quite low, and we
cannot allow the image set to remain on disk if there
is a failure.

In this patch we implement a flag to inform the snapshot
export, upload, download and import processes that it
should cleanup the snapshot path when it is done using
the data.

Some variables are also renamed to make them more
explicitly associated with this process.

ref: UPG-1518
Change-Id: I78a804a857ab15084853f0a4d512dfd1857bd4dd