[virsh] Implement snapshot flag for CI environments
In CI environments disk space can be quite low, and we
cannot allow the image set to remain on disk if there
is a failure.

In this patch we implement a flag to inform the snapshot
export, upload, download and import processes that it is
a CI environment and that the image set should be wiped
out for any failures. If there are no failures, the qcow
disks will be moved (not copied) when importing so that
there is no double disk space usage.

ref: UPG-1518
Change-Id: I78a804a857ab15084853f0a4d512dfd1857bd4dd