[virsh] Use ipaddr filter for address from domifaddr output
Some VM's may have multiple addresses, and the bash substitution
only removes the /prefix from the end of the whole multiline
string.

To make this more robust, we make use of the ipaddr filter to
extract just the valid addresses from the output and then only
select the first address for the connection test.

Change-Id: I5a28d31923912239352ff6f0f0a8019fdcded744