overcloud_reboot: Check if ceph-mon is located at ceph node
There are some topologies such as HCI where ceph-mon is not located
at controller nodes but colocated with other ceph services at ceph
nodes, in generally ceph mon could be located anywhere if composable
roles are used for deploying the overcloud. But we can assume for now
that It's either by default at controller nodes or It may be located
at the ceph node itself. This could be a quick and resonable fix.
A solution which could solve general case would need to be much more
robust and would require much more work and refactor.

Change-Id: I555603130b5e27ea767cdf9d1558ea5bdd81d842