Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> But if the SELinux binding is not provided, shouldn't something fail on the way?

Only if it SELinux is actually required. If it not required and and we can't provide it, we shouldn't care about not being able installing Python bindings for SELinux. I think actually this code we are executing in 'setup.py' should be better moved to an Ansible role IR plugins can use to require it.

Patch-set: 4