[virsh] On snapshot import, wait until the system has booted
We're seeing that later playbooks fail to gather facts afterwards
because SSH is not running. This may be due to the system boot
not yet being complete causing a flap in the SSH service.

To counter this, we wait until the system boot is complete before
finishing this play.

Previously we tried to use 'virsh domifaddr' to determine the right
addres to communicate to, but that was not reliable. The addresses
it found did not always allow SSH connections. Rather than use that
method we now use the infrared inventory details to get both the
right address and the right ssh user to use.

Change-Id: I4b0937def9a39c7c36fa2c4dee63288c04725023