Add release option into cdn_registery.
Now that the OSP versions are pinned to a specific rhel
version, when using Satellite we need to pin them at
subscription time. In Ansible 2.8, the rehat_subscription
module added a release option [0], however as Infrared
doesn't use Ansible 2.8 yet we can make use of it.

This patch uses the rhel_release option from cdn_creds to pin
the rhel version and removes it from the
redhat_subscription invoke, as infrared doesn't support
it. Once Ansible 2.8 is available in infrared we can just
move back this cdn_cred.rhel_release option.

NOTE: Changing the name from release to rhel_release as the
whole yaml is passed as an extra-environment file in infrared
and the release name is already being taken by rhos-release
role.

Change-Id: I35da84a98e376fcd364113b4a46054a255b93358