Add release option into cdn_registery.
Now that the OSP versions are pinned to a specific rhel
version, when using Satellite we need to pin them at
subscription time. In Ansible 2.8, the rehat_subscription
module added a release option [0], however as Infrared
doesn't use Ansible 2.8 yet we can make use of it.

This patch adds a new cdn_creds option named release
which when passed, will pin the rhel version. Like this,
once Ansible 2.8 is enabled in infrared we can just pass
this cdn_cred.release option as one of the redhat_
subscription options.

Change-Id: I35da84a98e376fcd364113b4a46054a255b93358