[tripleo-overcloud] Increase reboot timeout
Currently the reboot timeout is not enough. The reason for this is
that there are some systemd jobs which are waiting:

JOB UNIT                                 TYPE  STATE
295 systemd-update-utmp-runlevel.service start waiting
105 multi-user.target                    start waiting
297 openstack-nova-compute.service       start running
261 systemd-readahead-done.timer         start waiting

This may be related to the failing ovirt-guest-agent:

UNIT                      LOAD   ACTIVE SUB    DESCRIPTION
ovirt-guest-agent.service loaded failed failed oVirt Guest Agent

I have no idea why that guest agent is there, but to ensure that
systemd has time to complete its startup, we double the timeout.

We also remove the unnecessary task defaults for the tasks. Given
that the plugin has a default set, the value will always be defined.
This eliminates the duplication of timeout settings.

RHOSINFRA-3419
RHOSINFRA-3458

Change-Id: I621aad90db091178e9d0a70d01498d2bb5409736