[virsh] Remove unnecessary ovirt agent
The ovirt agent comes in the RHEL7.7 (and perhaps other)
virt images. However, if there is no ovirt device then
the agent service fails - but it takes a bit of time to
get to a failure point. That extra time slows down the
boot process significantly enough to be painful in CI.

We do not need the agent, so we simply remove it if it
is there.

RHOSINFRA-3419
RHOSINFRA-3458

Change-Id: I47411734e71bcd520b509724fb71bca1c7827cc3