Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I wonder , does this issue also has a BZ ?

I've reviewed these two:
https://bugzilla.redhat.com/show_bug.cgi?id=1061139
https://bugzilla.redhat.com/show_bug.cgi?id=1619665

Those explain why the package is installed in the first place, and for some reason there was a patch in later versions which make it auto-start.

If you have any suggestions about how best to deal with this, please let me know. This patch is really a workaround to improve OSP CI, but it doesn't solve the root issue of including a useless agent in the image.

Patch-set: 2