plugins: drop redundant cwd/chdir when git cloning
When git cloning infrared plugin, there is python approach used now,
where in past was direct git clone call via shell.
And in past for it to control where it should placed cloned content, changing cwd was used.
Now with python 'git.Repo.clone_from' approach this is unnecessary
as it is controlled by 'to_path' parameter.

But the presence of getcwd/chdir calls together with 'tenacity' based
retries and exception throwing can cause rare issue,
when if first fails it stays in tmpdir which it deletes before throwing
exception, causing then next retry to fail on getcw() as it is in now
deleted directory.

Simple solution is to just drop the now redunant getcwd/chdir calls
completely.

RHOSINFRA-3315

Original error appears as (shortened):

  WARNING Plugin 'tripleo-upgrade' has been successfully installed
  WARNING: Plugin 'tripleo-upgrade' has been successfully installed
  Traceback (most recent call last):
    File "/tmp/ir-venv-7kKVBVF/bin/infrared", line 8, in <module>
      sys.exit(main())
  ...
    File "/tmp/ir-venv-7kKVBVF/lib/python2.7/site-packages/infrared/core/services/plugins.py", line 455, in add_plugin
      self._clone_git_plugin(plugin_source, rev)
    File "/tmp/ir-venv-7kKVBVF/lib/python2.7/site-packages/tenacity/__init__.py", line 329, in wrapped_f
  ...
    File "/tmp/ir-venv-7kKVBVF/lib/python2.7/site-packages/tenacity/__init__.py", line 186, in reraise
      raise self.last_attempt.result()
  ...
    File "/tmp/ir-venv-7kKVBVF/lib/python2.7/site-packages/infrared/core/services/plugins.py", line 286, in _clone_git_plugin
      cwd = getcwd()
  OSError: [Errno 2] No such file or directory

Change-Id: I581ea8b396625c63b9e854fee2f49684c4cbdbcf