[virsh] - Add remove bridge step to cleanup
When virsh cleanup is performed, the configuration of the existing
brdige is deleted and the interface original configuration restored, but
the bridge remain as well as attached interface.

This could lead to an overcloud failure on connectivity verification in
case I'm using a different network topology in my next job.
Add a task to completely remove the bridge and attached interface.

Change-Id: I1391437abc0e082d0847f58d9c37eafe8082eaa4