Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> The RFE tells about adding TLS-E to OSP16.1, why IPAClient is been added to >=13 ?

TLS-E has been supported since OSP 13, but was only being tested in a handful of roles.
Upstream, all the templates in THT have (or should have) IPAClient defined since Queens.

TLS-E as implemented in 16.1 reuses this template, but points it to a different composable service.  

By default, IPAClient is set to OS::None, so there is no issue in breaking existing tests.

Patch-set: 3
CC: Gerrit User 1006809 <1006809@d5d70762-12d0-45a1-890d-524b12d3f735>