[tempest] use the correct requirements file
test-requirements.txt contains the requirements used to test
the repository itself, not the requirements needed to use
that repository.
This holds even when the repository itself
contains tests and test frameworks: requirements.txt
lists the requirements needed to install the test code,
while test-requirements.txt contains the tests for the tests
(mostly linters, unit tests when needed, etc).

Change-Id: I66cf3e3b8174a0ac059710cd74caa861b756aaa2