Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> *main flows (gates, IR p1/p2 jobs, nigthly, ...)

The purpose of the code freeze is to protect the shared code and the default behavior. As this only affects a special, non-default case, I believe the freeze doesn't apply, or at least the change is not against the spirit of the freeze, if it can unblock people.

Patch-set: 7
Reviewer: Gerrit User 1016110 <1016110@d5d70762-12d0-45a1-890d-524b12d3f735>