[tempest] Converge to using the common create_venv role for git-based installs
Rather than duplicate this code, we converge the role's git installation to
use the common create_venv role. In order to achieve this, we need to add the
capability to the create_venv role to be able to install from a requirements
files.

This ensures that the installation is done in a consistent way with the other
infrared plugins which improves reliability.

RHOSINFRA-3168

Change-Id: I4dc74d52adc50e3d040fd8ff5706855481fe2c92