Modify the way we enable extra repositories
Today, the task for downloading repositories will work only if the
provided URL is of a file that ends with .repo

We would like to also support a use case of providing a URL to
the repository itself (the path which holds the packages).

The way it can be done is with dnf/yum
config-manager --add-repo which supports both use cases.

Change-Id: I0ec0668b202574375693e1d18b9918b1593330a2