[tempest] Use the common create_venv role for pip-based installs
Rather than duplicate this code, we converge the role's pip installation
to use the common create_venv role.

This ensures that the installation is done in a consistent way with the
other infrared plugins which improves reliability.

To ensure that this works in a similar way to the git-based installation,
we retain the lack of a version cap for the 'tempest' package and we
respect the use of the force_pyvenv variable.

We are able to remove unnecessary packages from setup.packages because
the create_venv role will handle the required package installation where
needed. We also sort the setup.pip_packages alphanumerically.

RHOSINFRA-3168

Change-Id: If0b8b2df570cac945048ed1672acbfddc211690c