[collect-logs] Find correct python interpreter
If /usr/bin/python is not available when trying to collect
the logs, then none are collected. In this patch we ensure
that alternative python interpreters are considered when
attempting the log collection.

We also correct the formatting for the other playbook vars
which should be a dict, not a list. The list format is for
vars_files, not vars.

RHOSINFRA-3520

Change-Id: Iab70ceed39c03e75f7daa5143ce4c943581f1ff9