[virsh] Use the ssh key from the imported workspace
We do not need to inject the host's key into the image because we
save the key that was used on the build host.

Also, when we check to see if the VM is up, we should use the right
key from the imported workspace. We also have to delegate this task
to the ansible controller which is where the workspace is.

UPG-1718
UPG-1725

Change-Id: I00fff1732d340223fa10a2ebc49cc0c924665871