[collect-logs] Add a default python interpreter for unreachable hosts
When a host is unreachable, currently the python interpreter variable
is not resolvable. We have two options to solve this as far as I can
see:

1. Provide a default in the var so that it is always resolvable.
2. Use set_fact to only set the interpreter on reachable hosts.

In this patch we opt to do the first and set it to the only interpreter
that is guaranteed to be there (because the system uses it).

RHOSINFRA-3520

Change-Id: I54c3f1c7b2c9a9649f0a706eaa40ec78f2f05956