[virsh] Create/update /etc/hosts entries when importing
When the image set import happens, it is useful to also ensure
that the /etc/hosts file includes the reference to the VM name
and its address. To make this useful, we also ensure that the
hypervisor's public SSH key is added to authorized_keys in all
running VM's.

UPG-1710

Change-Id: I6a1408cdc4b18a0ae351a2f256e9d63733532689