[virsh] Ensure an image set exists before downloading
Currently if the image set does not exist, then the download
by the swift client returns success and the process continues,
resulting in an obscure failure further along the process.

In this patch we first validate that the image set exists, then
download it. If it does not exist, we fail the task immediately.

UPG-1569

Change-Id: I324539549d680cb037d56c41cc14dc5ea7b3460b