Remove env dependent template guess
undercloud bahaved differently when hypervisor or bmc
node was defined. It assumed (correctly) it should try to
use the in repo undercloud template.

However it should be default behavior even when the hypervisor or bmc
node is hidden from the undercloud.

Technically passing 'none' was a working as I updated the help text,
however unlikely anybody really wants to use the sample undercloud
nowadays.
The default behavior will be to use the built-in template,
this is the only real change.

Change-Id: I3503cc748a0785580e28d9a4e8e455f4562a3ca9