Update patch set 1
Patch Set 1:

Luca, is this breaking in rhos13 deploys? the reason i ask is ipaservices-baremetal-ansible.yaml.j2 should only be getting run with "--tls-everywhere yes --enable-novajoin no" is passed.  --enable-novajoin yes is the default. This would only be run with 16.1 when using the new enrollment method and its not going to work in 16 or lower releases. 

See:
https://github.com/redhat-openstack/infrared/blob/master/infrared/common/roles/tls-everywhere/tasks/prepare_overcloud.yml#L57

There is  MakeHomeDir, IdMDomain, and IdMNoNtpSetup that can be used for parameter_defaults in older releases but you should not need to reference /extraconfig/services/ipaclient.yaml. The other defaults referenced would not get used at all as they are not in the old template. 
https://code.engineering.redhat.com/gerrit/gitweb?p=openstack-tripleo-heat-templates.git;a=blob;f=extraconfig/services/ipaclient.yaml;h=3267569e80cffb0bfcc82f660765544054733aff;hb=refs/heads/rhos-13.0-patches

Patch-set: 1
Reviewer: Gerrit User 1015419 <1015419@d5d70762-12d0-45a1-890d-524b12d3f735>