[virsh] workaround for bz1840307
https://bugzilla.redhat.com/show_bug.cgi?id=1840307

Restart the VM if we can't get IP for it.

Update (fhubik): single cycle of ^ still hits this issue,
but in only in 2 of 30 cases (based on 9/3O without any fix).
Repeating of problematic steps mult. (shorter) times (3)
tries to lower chance of hitting it to zero (so there
are no jobs randomly failing on this at all).

RHOSINFRA-3266

Change-Id: I637ebdbbd598e5cf09823670f232dc32d3bcbdfa