Remove tag discovery option
"install.registry.tag" Cannot be a string and
a dict with a 'discovery' option at the same time.
As far as I can tell the option would simply switch the
tag value to AUTODETECT which could be done by passing
autodetect to install.registry.tag instead of having a new option.

Change-Id: I09f67d61f00d6997284bcecb2ea8277d56e98084