Overcloud do not expect old args
registry-skip-puddle assumes cdn usage however it is not the
only case when you may skip puddle.
The option originally used only to try avoid rhos-release usage
or depending on it's side effects.

Since both the undercloud and overcloud containerized,
we can simple have overcloud to use the underclouds registry
without insisting on passing any extra override options, so
this change allows to not pass unnecessary arguments, if the
user only intention is to skip puddle related tool usage.

Change-Id: I3684d3a8d5deb93ee25b939a1983c07b74ea7c9a