Overcloud do not expecet old args
registry-skip-puddle assumes cdn usage however it is not the
only case when you may skip puddle.
The option orginally used only to try avoid rhos-release usage
or depending on it's side effects.

Since both the undercloud and overcloud contenarized,
we can simple have overcloud to use the underclouds registry
without insisting on extra override options, so
allowing not in use options to not beeing passed.

Change-Id: I3684d3a8d5deb93ee25b939a1983c07b74ea7c9a