Use Python3 for Octavia / Tempest in OSP 13.0
Octavia Tempest plugin dropped support for Python2 and now
there are tests that require Python3. Therefore we want to
start running Tempest tests against Python3. For this, we
need to adjust venv_distro_package_list on OSP13 and when
the Python3 is forced in venv (by default the create_venv
role would attempt to install python3-devel package, which
is not present in RHEL7). Additionally the venv_pip_packages
variable is cleaned from empty element as it causes failure
when used in old call to pip version in RHEL7.

Jira: OSPNET-39
Change-Id: I313d3a8d0947980282d5b4dba73e49267528612a