Update patch set 4
Patch Set 4:

I think people use tempest from git in this case, which is already supported and doesn't require per-component hacks (which is not documented in the first line of the commit message). 
Doesn't it work for you?

Alternative solution (RPM-based): make sure your tempest tests are still working with python 2 and backport the package in RDO. Manila and Cinder followed this way for their respective tempest plugins.

Patch-set: 4
CC: Gerrit User 1016110 <1016110@d5d70762-12d0-45a1-890d-524b12d3f735>