Add ability to supply custom standalone parameters
Commit adds the ability to provide custom standalone files and
deployment script when creating a standalone deployment. This will allow
end users the capacity to leverage different features when deploying
standalone openstack, such as deploying with SR-IOV, AMD Sev, or GPU.

The approach takes what is currently used in tripleo-overcloud when
handling custom deployment scripts and directories [1]. It also
introduces several new command line parameters to the plugin.
deployment-script takes the template or script that will be used to
deploy the standalone stack. deployment-files is the directory path
containing template files or normal files with the deployment
parameters. Lastly deploy-ip allows the end user to provide their own IP
address to be used by the deployment if they do not wish to use what is
provided on the local-interface. If none of these parameters are
provided in the cli, the deployment will default to the standard script
and parameters for deploying standalone.

[1] https://github.com/redhat-openstack/infrared/blob/master/plugins/tripleo-overcloud/build_templates.yml

Change-Id: I9684d356cfb033eea6c6b1e76a94a3504ca68f82