Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I think people use tempest from git in this case, which is already supported and doesn't require per-component hacks (which is not documented in the first line of the commit message). 
> Doesn't it work for you?
> 
> Alternative solution (RPM-based): make sure your tempest tests are still working with python 2 and backport the package in RDO. Manila and Cinder followed this way for their respective tempest plugins.

Sorry, the commit summary is a bit misleading probably, I will update it – it is not about the Tempest itself, but Octavia-Tempest-plugin is the guilty guy here, as it started to contain Python3-only code. So, we were reported that we would like to follow execution with Python3. I am afraid hacks will be necessary here in such case.

Patch-set: 4