Prefer IPv4 over IPv6 on undercloud
By default IPv6 has preference over IPv4 and in case there are
issues with external IPv6 connectivity it causes long delays
e.g. there is a 10 minutes delay on introspection stage when
trying to download upper constrains).
This patch adds a flag to prefer IPv4 over IPv6 on the undercloud
(the flag is enabled by default).

Change-Id: I8d79805dcc6399b8515d5d24bf274cb537ce037a