Replace selinux_fix hack with selinux shim
The selinux_fix hack is fragile and local to this project.
The fragility is due to changing locations for the selinux
python bindings, or changes in the underlying kernel, or
the supporting binaries.

Rather than do the binding copy into the virtualenv, we can
use this shim to just load them from the system instead. The
shim is a widely available tool, and therefore not specific
to this project. This makes it more likely to be better
maintained. It also happens that the author/maintainer works
at RedHat.

RHOSINFRA-3660

Change-Id: I4434e1a41dbfdec8b976621c795b654da7ea129f