Update patch set 2
Patch Set 2:

Just a few of things. First it does not look like I have any permissions to do +V for this commit. Second I have tested the dynamic and standard deployments, but noticed an issue with standard deployments unrelated to my commit. This looks to be a problem with the current standard template for parameters, specifically [1]. This is creating container tags that the deployment cannot find during deployment. Downstream this deployment is done with a variant of the plugin that ignores setting this parameter [2] entirely. Not sure if I should do the same as [2] in this commit, make the change in another commit, or additional logic needs to be added if that parameter is needed in certain RHOS versions.

Lastly it seems like this plugin has gotten some activity recently. If its use case is going to see increased activity going forward maybe we can discuss the idea of leveraging the DF DFG's standalone job for CI verification of standalone plugin changes?

[1] https://github.com/redhat-openstack/infrared/blob/master/plugins/tripleo-standalone/templates/standalone_parameters.yaml.j2#L12

[2] https://code.engineering.redhat.com/gerrit/gitweb?p=rhos-qe-core-installer.git;a=blob;f=ansible_roles/standalone_deployment/templates/standalone_parameters.yaml.j2;hb=HEAD#l12

Patch-set: 2