Introspection: do not poweroff custom oc_ironic nodes
Custom OC Ironic nodes are not part of regular OC
nodes supposed to be introspected/provisioned as part
of overcloud deploy - they are instead added post-deployment
into OC Ironic as bm nodes to manage.

And as such they do not be prepped for introspection,
so no need to power them off.

RHOSINFRA-3689

Change-Id: I396f9e2397b1b80aa26130089369207f8c76b675